DS | 02/09/2023 10:40 # Totalling Balance, Income and Expanses And also totalling by datewise

+++ lib/modals/transaction_modal.dart
@@ -5,7 +5,7 @@ class TransactionModal {
   String? remarks;
   String? type;
   String? category;
-  double? amount;
+  int? amount;
   String? date;
   String? time;
